refactor examples palcement
